Combine to one file
