+++ requirements.txt
@@ -1,7 +1,6 @@
 fastapi
 uvicorn[standard]
 beanie
-motor
 bcrypt
 python-dotenv
-email-validator
+pydantic-email
